Move `@maybe_return_call` back to `DimensionAnalysis.jl`
